Remove reassignment of body variable
I had previously changed the function to not modify any objects that are passed in as function parameters.
But actually that is exactly what we need. We need to change the bodyElement that is passed into the function.

To fix the error that occured whenever the function was triggered, I had to remove the reassignment.