Fix key is not unique error in SimpleNavigation
It seems, that using the full <Link/> object as a key was not unique,
so I switched it to using the map index as the key for the NavigationLinks.

This is not best practice, but because we preload the naviation links, the order will
not change at any point, so I think it's safe to use.