Fix PageInternalLink ignoring onClick function
The issue was that the PageInternalLink srolls to the hash on click. This meant we ignored any other onClick props.
We now execute the onClick passed as a prop after we execute the srollToHash function.
We now also pass through any other props that might be passed to the PageInternalLink